introduce multiplexers in separate step, avoiding unnecessary ones with input-independent nodes
